backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "scikit-image>=0.25.2",
    "uvicorn>=0.35.0",
]
